Fix problem with the collection field
I removed the ability to set a key with the collection
field since that needs to be an index. This means if
somebody calls `field` from within a `collection` block with
a name, it will give an error.

Thats kind of confusing, so consider removing this method, making
the error message clearer, or rationalizing this better.