Well this is fun.
In my controller I have one view. Each controller action corresponds to
a state change in the view. It changes the state of an object and renders
the view, which makes it look seamless.

This is for batch manipulations.

I know this isn't new ... I've seen this done this way in regular HTML
back in the day for awesome things like panning around map tiles. What
could make this interesting for Rails is incorporating Turbo, or maybe
a lower level DOM diffing JS.

The way the code laid out feels interesting. Right now its specific to
button actions, but its not hard to imagine how state transitions between
them could be handled by a Phlex view and on the controller.